CONTENT-8440: add 0.5 option for MessagingThrottlingInterval setting.

+++ src/Modules/Settings/Dnn.PersonaBar.SiteSettings/Services/Dto/UpdateMessagingSettings.cs
@@ -34,7 +34,7 @@ namespace Dnn.PersonaBar.SiteSettings.Services.Dto
 
         public bool DisablePrivateMessage { get; set; }
 
-        public int ThrottlingInterval { get; set; }
+        public double ThrottlingInterval { get; set; }
 
         public int RecipientLimit { get; set; }
